AIHackOption1/tender_evaluation_system/requirements.txt
psycopg2-binary
jinja2
pandas
numpy
matplotlib
seaborn
reportlab
docxtpl
